# project-service/src/main/java/com/saas/project_service/entity/Project.java
package com.saas.project_service.entity;

import jakarta.persistence.*;
import lombok.*;
import org.hibernate.annotations.CreationTimestamp;
import org.hibernate.annotations.UpdateTimestamp;

import java.time.Instant;
import java.time.LocalDate;
import java.util.UUID;

@Entity
@Table(name = "projects")
@Getter
@Setter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class Project {

    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    private UUID id;

    @Column(name = "workspace_id", nullable = false)
    private UUID workspaceId;

    @Column(nullable = false)
    private String name;

    @Column(columnDefinition = "TEXT")
    private String description;

    @Enumerated(EnumType.STRING)
    @Column(nullable = false, length = 30)
    @Builder.Default
    private ProjectStatus status = ProjectStatus.PLANNING;

    @Column(name = "client_visible", nullable = false)
    @Builder.Default
    private boolean clientVisible = true;

    @Column(name = "start_date")
    private LocalDate startDate;

    @Column(name = "target_date")
    private LocalDate targetDate;

    @Column(name = "progress_percent", nullable = false)
    @Builder.Default
    private int progressPercent = 0;

    @Column(name = "demo_url", length = 1000)
    private String demoUrl;

    @Column(name = "github_repo_owner")
    private String githubRepoOwner;

    @Column(name = "github_repo_name")
    private String githubRepoName;

    @Column(name = "github_repo_branch")
    private String githubRepoBranch;

    /** Pemilik token GitHub untuk repo ini. Commit dibaca dengan token orang ini,
     *  bukan token penonton — CLIENT tidak punya koneksi GitHub sendiri. */
    @Column(name = "github_linked_by")
    private UUID githubLinkedBy;

    @Column(name = "created_by", nullable = false)
    private UUID createdBy;

    @CreationTimestamp
    @Column(name = "created_at", nullable = false, updatable = false)
    private Instant createdAt;

    @UpdateTimestamp
    @Column(name = "updated_at", nullable = false)
    private Instant updatedAt;

    @Column(name = "archived_at")
    private Instant archivedAt;
}